Fixed some changes inlcuding putting FMS calculations out of figure files.
